ch: Handle resize events
When resize events are found in the monitor, try to update the thread
id mapping (if a change is detected the domain.xml will also be
updated).

On vm resizes removing vcpus, the event will be triggered but because
cloud-hypervisor does not kill the associated thread right away, the
thread of the removed vcpu is going to still be detected and so
validating the event was processed successfully is skipped for now.