Enable tests to run in parallel
Adds bubblewrap to the test system dependencies in order to mount
various content into the temporary namespace and continue to run as
the current user.

Each test then is run individually in the namespace with a private
/run which is the key requirement to allow libvirt to run in
parallel. The use of /home and /tmp seems to be sharable without
running into conflicts due to the names of the vms in libvirt being
distinct.

Signed-off-by: William Douglas <william.douglas@intel.com>