ch_monitor: Make unused function static
The virCHMonitorGet function wasn't in use and was declared as
non-static (though not in a header file). This function isn't going to
be used outside of the monitor though so remove the initial
declaration and define the function to be static.

Future work should adjust this function to allow reading VM state
needed for PTY enablement.

Signed-off-by: William Douglas <william.douglas@intel.com>