Always end job in SrcBegin
This is what QEMU does when CHANGE_PROTECTION is not implemented. The
downside of course is someone else can come in between to change the
domain, but implementing the more complex locking scheme before we have
working migration support seems to be barking at the wrong tree.

Drop EndJob in other places. They were paired with the one in SrcBegin.
They aren't needed anymore.

Signed-off-by: Wei Liu <liuwe@microsoft.com>